FirstStepsREADME.md
Knowledge is key as long as it's used In the proper way. Life is filled with trials and tribulations. Not sure if this really matters. I'm not scared to lose but will gain from first hand user experience. Testing the material from start to exit to start to exit again.